+++ rex/utils/wrapper.py
@@ -11,6 +11,8 @@ import traceback
 from email.mime.text import MIMEText
 from typing import Callable, Optional
 
+from accelerate.state import AcceleratorState
+
 from rex.utils.logging import logger